Added Console bindings to the UI.
There is now a key handler added to the application, it currently listens for a user pressing enter. And will do nothing with it until the Console class is added.
If you want to add a new key, go to the handle jey event and add it to the switch statement inside.